+++ src/js/draw.js
@@ -81,9 +81,21 @@ var drawutils = /** @class */ (function () {
      *
      * @name setCurrentId
      * @method
-     * @param {UID=} uid - (optional) A UID identifying the currently drawn element(s).
+     * @param {UID} uid - A UID identifying the currently drawn element(s).
      **/
     drawutils.prototype.setCurrentId = function (uid) {
+        // NOOP
+    };
+    ;
+    /**
+     * This method shouled be called each time the currently drawn `Drawable` changes.
+     * Determine the class name for further usage here.
+     *
+     * @name setCurrentClassName
+     * @method
+     * @param {string} className - A class name for further custom use cases.
+     **/
+    drawutils.prototype.setCurrentClassName = function (className) {
         // NOOP
     };
     ;