+++ src/js/drawgl.d.ts
@@ -62,9 +62,18 @@ export declare class drawutilsgl implements DrawLib<void> {
      *
      * @name setCurrentId
      * @method
-     * @param {UID=} uid - (optional) A UID identifying the currently drawn element(s).
+     * @param {UID} uid - A UID identifying the currently drawn element(s).es.
      **/
-    setCurrentId(uid?: UID): void;
+    setCurrentId(uid: UID | undefined): void;
+    /**
+     * This method shouled be called each time the currently drawn `Drawable` changes.
+     * Determine the class name for further usage here.
+     *
+     * @name setCurrentClassName
+     * @method
+     * @param {string} className - A class name for further custom use cases.
+     **/
+    setCurrentClassName(className: string | undefined): void;
     /**
      * Draw the line between the given two points with the specified (CSS-) color.
      *