+++ src/js/interfaces.js
@@ -3,6 +3,8 @@
  * @author Ikaros Kappler
  * @modified 2021-01-10 Added the `CanvasWrapper` interface.
  * @modified 2021-01-20 Added the `UID` type.
+ * @modified 2021-01-25 Added the `DrawLib.setCurrentId` and `DrawLib.setCurrentClassName` functions.
+ * @modified 2021-01-25 Fixed the `PBParams` interface (inluding DrawConfig).
  **/
 Object.defineProperty(exports, "__esModule", { value: true });
 //# sourceMappingURL=interfaces.js.map