+++ src/ts/PlotBoilerplate.ts
@@ -511,11 +511,12 @@ export class PlotBoilerplate {
 		throw `The svg draw library is not yet integrated part of PlotBoilerplate. Please include ./src/js/utils/helpers/drawutils.svg into your document.`;
 	    this.canvas = canvasElement as SVGElement;
 	    this.draw = new drawutilssvg( this.canvas as SVGElement,
-					  new Vertex(), // offset
-					  new Vertex(), // scale
+					  new Vertex(),    // offset
+					  new Vertex(),    // scale
 					  this.canvasSize,
-					  false,        // fillShapes=false
-					  true          // isPrimary=true
+					  false,           // fillShapes=false
+					  this.drawConfig,
+					  true             // isPrimary=true
 					);
 	    this.fill = (this.draw as drawutilssvg).copyInstance( true ); // fillShapes=true
 	    
@@ -986,6 +987,8 @@ export class PlotBoilerplate {
 	    var d : Drawable = this.drawables[i];
 	    this.draw.setCurrentId(d.uid);
 	    this.fill.setCurrentId(d.uid);
+	    this.draw.setCurrentClassName(d.className);
+	    this.draw.setCurrentClassName(d.className);
 	    this.drawDrawable( d, renderTime, draw, fill );
 	}
     };
@@ -1015,6 +1018,7 @@ export class PlotBoilerplate {
 		    if( !d.bezierCurves[c].startPoint.attr.bezierAutoAdjust ) {
 			if( d.bezierCurves[c].startPoint.attr.visible ) {
 			    draw.setCurrentId(`${d.uid}_h0`);
+			    draw.setCurrentClassName(`${d.className}-start-handle`);
 			    draw.diamondHandle( d.bezierCurves[c].startPoint, 7, this._handleColor(d.bezierCurves[c].startPoint,this.drawConfig.vertex.color) );
 			}
 			d.bezierCurves[c].startPoint.attr.renderTime = renderTime;
@@ -1022,16 +1026,19 @@ export class PlotBoilerplate {
 		    if( !d.bezierCurves[c].endPoint.attr.bezierAutoAdjust ) {
 			if( d.bezierCurves[c].endPoint.attr.visible ) {
 			    draw.setCurrentId(`${d.uid}_h1`);
+			    draw.setCurrentClassName(`${d.className}-end-handle`);
 			    draw.diamondHandle( d.bezierCurves[c].endPoint, 7, this._handleColor(d.bezierCurves[c].endPoint,this.drawConfig.vertex.color) );
 			}
 			d.bezierCurves[c].endPoint.attr.renderTime = renderTime;
 		    }
 		    if( d.bezierCurves[c].startControlPoint.attr.visible ) {
 			draw.setCurrentId(`${d.uid}_h2`);
+			draw.setCurrentClassName(`${d.className}-start-control-handle`);
 			draw.circleHandle( d.bezierCurves[c].startControlPoint, 3, this._handleColor(d.bezierCurves[c].startControlPoint,'#008888') );
 		    }
 		    if( d.bezierCurves[c].endControlPoint.attr.visible ) {
 			draw.setCurrentId(`${d.uid}_h3`);
+			draw.setCurrentClassName(`${d.className}-end-control-handle`);
 			draw.circleHandle( d.bezierCurves[c].endControlPoint, 3, this._handleColor(d.bezierCurves[c].endControlPoint,'#008888') );
 		    }
 		    d.bezierCurves[c].startControlPoint.attr.renderTime = renderTime;
@@ -1044,7 +1051,11 @@ export class PlotBoilerplate {
 		}
 		
 		if( this.drawConfig.drawBezierHandleLines && this.drawConfig.drawHandleLines ) {
+		    draw.setCurrentId(`${d.uid}_l0`);
+		    draw.setCurrentClassName(`${d.className}-start-line`);
 		    draw.line( d.bezierCurves[c].startPoint, d.bezierCurves[c].startControlPoint, this.drawConfig.bezier.handleLine.color, this.drawConfig.bezier.handleLine.lineWidth );
+		    draw.setCurrentId(`${d.uid}_l1`);
+		    draw.setCurrentClassName(`${d.className}-end-line`);
 		    draw.line( d.bezierCurves[c].endPoint, d.bezierCurves[c].endControlPoint, this.drawConfig.bezier.handleLine.color, this.drawConfig.bezier.handleLine.lineWidth );
 		}
 		
@@ -1063,11 +1074,14 @@ export class PlotBoilerplate {
 	} else if( d instanceof VEllipse ) {
 	    if( this.drawConfig.drawHandleLines ) {
 		draw.setCurrentId(`${d.uid}_e0`);
+		draw.setCurrentClassName(`${d.className}-v-line`);
 		draw.line( d.center.clone().add(0,d.axis.y-d.center.y), d.axis, '#c8c8c8' );
 		draw.setCurrentId(`${d.uid}_e1`);
+		draw.setCurrentClassName(`${d.className}-h-line`);
 		draw.line( d.center.clone().add(d.axis.x-d.center.x,0), d.axis, '#c8c8c8' );
 	    }
-	    draw.setCurrentId(d.uid); 
+	    draw.setCurrentId(d.uid);
+	    draw.setCurrentClassName(`${d.className}`);
 	    draw.ellipse( d.center, Math.abs(d.axis.x-d.center.x), Math.abs(d.axis.y-d.center.y), this.drawConfig.ellipse.color,  this.drawConfig.ellipse.lineWidth );
 	    if( !this.drawConfig.drawHandlePoints ) {
 		d.center.attr.renderTime = renderTime;
@@ -1080,7 +1094,7 @@ export class PlotBoilerplate {
 	} else if( d instanceof Vertex ) {
 	    if( this.drawConfig.drawVertices &&
 		(!d.attr.selectable || !d.attr.draggable) && d.attr.visible ) {
-		// Draw as special point (grey)
+		// Draw as special point (grey)		
 		draw.circleHandle( d, 7, this.drawConfig.vertex.color );
 		d.attr.renderTime = renderTime;
 	    }
@@ -1094,6 +1108,7 @@ export class PlotBoilerplate {
 	    draw.arrow( d.a, d.b, this.drawConfig.vector.color );
 	    if( this.drawConfig.drawHandlePoints && d.b.attr.selectable && d.b.attr.visible ) {
 		draw.setCurrentId(`${d.uid}_h0`);
+		draw.setCurrentClassName(`${d.className}-handle`);
 		draw.circleHandle( d.b, 3, '#a8a8a8' );
 	    } else {
 		d.b.attr.renderTime = renderTime;	
@@ -1106,12 +1121,14 @@ export class PlotBoilerplate {
 	} else if( d instanceof PBImage ) {
 	    if( this.drawConfig.drawHandleLines ) {
 		draw.setCurrentId(`${d.uid}_l0`);
+		draw.setCurrentClassName(`${d.className}-line`);
 		draw.line( d.upperLeft, d.lowerRight, this.drawConfig.image.color, this.drawConfig.image.lineWidth );
 	    }
 	    fill.setCurrentId(d.uid);
 	    fill.image( d.image, d.upperLeft, d.lowerRight.clone().sub(d.upperLeft) );
 	    if( this.drawConfig.drawHandlePoints ) {
 		draw.setCurrentId(`${d.uid}_h0`);
+		draw.setCurrentClassName(`${d.className}-lower-right`);
 		draw.circleHandle( d.lowerRight, 3, this.drawConfig.image.color );
 		d.lowerRight.attr.renderTime = renderTime;
 	    }